l2-l1 interface refactoring
